+++ frontend/src/app/features/main-page/components/add-game-form/add-game-form.component.ts
@@ -76,7 +76,7 @@ export class AddGameFormComponent extends Destroyable implements OnInit {
                 })
             ).subscribe(() => {
                 this.viewState = ViewState.SUCCESS;
-                this.form.reset();
+                this.initForm();
                 this.reloadDescription.emit();
             });
     }
@@ -89,7 +89,7 @@ export class AddGameFormComponent extends Destroyable implements OnInit {
     private prepareAddGameDataForRequest(): AddGameRequestBody {
         const requestBody: AddGameRequestBody = {
             name: this.form.get(AddGameFormControl.NAME)?.value,
-            hasNewAchievements: this.form.get(AddGameFormControl.NEW_ACHIEVEMENTS)?.value,
+            hasNewAchievements: this.form.get(AddGameFormControl.HAS_NEW_ACHIEVEMENTS)?.value,
             categories: [],
         };
 
@@ -113,7 +113,6 @@ export class AddGameFormComponent extends Destroyable implements OnInit {
         this.editGameService.editingGameName$.pipe(filter(Boolean), takeUntil(this.destroyed$)).subscribe((gameName: string) => this.updateFormWhenEditingGame(gameName));
     }
 
-    // TODO: Fix game edit
     private updateFormWhenEditingGame(gameName: string): void {
         this.form.reset();
 
@@ -121,36 +120,36 @@ export class AddGameFormComponent extends Destroyable implements OnInit {
         const game: Game = this.findGameByName(completedGames, gameName) || this.findGameByName(gamesWithNewAchievements, gameName);
 
         this.form.get(AddGameFormControl.NAME)?.setValue(gameName);
+        this.form.get(AddGameFormControl.HAS_NEW_ACHIEVEMENTS)?.setValue(game.hasNewAchievements);
         this.updateFormCategoriesWhenEditingGame(game);
         this.form.updateValueAndValidity();
         this.viewportScroller.scrollToAnchor(this.headingId);
     }
 
-    // TODO: Fix game edit
     private updateFormCategoriesWhenEditingGame(game: Game): void {
         game.categories?.forEach((gameCategory: GameCategory) => {
             if (gameCategory.type === CategoryType.LONG_GAME || gameCategory.type === CategoryType.VERY_LONG_GAME || gameCategory.type === CategoryType.ULTRA_LONG_GAME) {
-                this.form.get(AddGameFormControl.LENGTH)?.setValue(gameCategory.type);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.LENGTH)?.setValue(gameCategory.type);
             }
 
             if (gameCategory.type === CategoryType.HARD_GAME || gameCategory.type === CategoryType.VERY_HARD_GAME || gameCategory.type === CategoryType.ULTRA_HARD_GAME) {
-                this.form.get(AddGameFormControl.DIFFICULTY)?.setValue(gameCategory.type);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.DIFFICULTY)?.setValue(gameCategory.type);
             }
 
             if (gameCategory.type === CategoryType.LOVED_GAME) {
-                this.form.get(AddGameFormControl.LOVED_GAME)?.setValue(true);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.LOVED_GAME)?.setValue(true);
             }
 
             if (gameCategory.type === CategoryType.BAD_GAME) {
-                this.form.get(AddGameFormControl.BAD_GAME)?.setValue(true);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.BAD_GAME)?.setValue(true);
             }
 
             if (gameCategory.type === CategoryType.DOESNT_COUNT) {
-                this.form.get(AddGameFormControl.DOESNT_COUNT)?.setValue(true);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.DOESNT_COUNT)?.setValue(true);
             }
 
             if (gameCategory.type === CategoryType.BUGGED_GAME) {
-                this.form.get(AddGameFormControl.BUGGED_GAME)?.setValue(true);
+                this.gameCategoriesFormGroup.get(AddGameFormControl.BUGGED_GAME)?.setValue(true);
             }
         });
     }
@@ -162,7 +161,7 @@ export class AddGameFormComponent extends Destroyable implements OnInit {
     private initForm(): void {
         this.form = this.formBuilder.group({
             [AddGameFormControl.NAME]: ['', Validators.required],
-            [AddGameFormControl.NEW_ACHIEVEMENTS]: [false],
+            [AddGameFormControl.HAS_NEW_ACHIEVEMENTS]: [false],
             [AddGameFormGroup.GAME_CATEGORIES]: this.formBuilder.group({
                 [AddGameFormControl.LENGTH]: [null],
                 [AddGameFormControl.DIFFICULTY]: [null],